Learn: add a backward graph when parsing sentences
This will allow to build sentences by appending words both *before*
and after existing words.

Note that this commit changes the format of data stored on-disk, so it
breaks existing use of the plugin. A later commit will make this
change backward-compatible by silently converting the old format to
the new one.