Don't loop over faulty on_run events
When a plugin throws an exception inside its 'on_run' method, we were
not resetting the timer to the next scheduled call.

This leads the main code to call the 'on_run' method of the plugin
again and again, thus creating an infinite loop if the plugin keeps
throwing an exception.

Simply resetting the timer before actually calling the method solves
this issue.

This bug was spotted because the microblogging plugin throws an
uncatched exception.